requirements.txt
ragas==0.1.15
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
langchain==0.1.0
langchain-openai==0.0.5
